feat: 경로 검색 결과 UI 수정

+++ domain/src/main/java/com/stop/domain/model/route/tmap/custom/Route.kt
@@ -5,6 +5,6 @@ interface Route {
     val end: Place
     val mode: MoveType
     val sectionTime: Double
-    val proportionOfSectionTime: Int
+    val proportionOfSectionTime: Float
     val start: Place
 }
